catégorie matériel index refactoring

+++ src/app/views/categories/categorie-materiel-index/categorie-materiel-index.component.ts
@@ -1,38 +1,72 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnDestroy, OnInit, ViewChild} from '@angular/core';
 import { ToasterService } from 'angular2-toaster';
 import { CategorieMaterielService } from '../categorie-materiel.service';
+import {DataTableService} from '../../common/utils/data-table.service';
+import {DataTableDirective} from 'angular-datatables';
+import {Subject} from 'rxjs';
 
 @Component({
   selector: 'app-categorie-materiel-index',
   templateUrl: './categorie-materiel-index.component.html',
   styleUrls: ['./categorie-materiel-index.component.css']
 })
-export class CategorieMaterielIndexComponent implements OnInit {
-  categoriesMaterielsData = null;
-  constructor(private categorieMaterielService: CategorieMaterielService,
-    private toasterService: ToasterService) { }
+export class CategorieMaterielIndexComponent implements OnDestroy, OnInit {
+  // datatabale triggers
+  @ViewChild(DataTableDirective, {static: false})
+  dtElement: DataTableDirective;
+  // datatable data
+  categoriesMaterielsData: Array<any> = [];
+  // @ts-ignore
+  dtTrigger: Subject = new Subject();
+  // datatable options
+  dtOptions: DataTables.Settings = {};
 
-ngOnInit(): void {
-this.loadCategorieMateriel();
-}
+  constructor(
+    private categorieMaterielService: CategorieMaterielService,
+    private toasterService: ToasterService,
+    private dataTableService: DataTableService) { }
 
-// pour charger les catégories Materielfrom REST API
-loadCategorieMateriel() {
-this.categorieMaterielService.getAllCategorieMateriel().subscribe(
-(bodyResponse) => { this.getCategorieMaterielData(bodyResponse); });
-}
+  ngOnInit(): void {
+    this.dtOptions = this.dataTableService.getDataTableOptions();
+    this.loadCategorieMateriel();
+  }
 
-// get catégorie data from bodyresponse
-getCategorieMaterielData(bodyResponse) {
-this.categoriesMaterielsData = bodyResponse.data;
-}
-// delete catégorie
-deleteCategorieMateriel (id) {
-this.categorieMaterielService.deleteCategorieMateriel(id).subscribe(
-(bodyResponse) => { this.getMessage(bodyResponse); });
-}
+  // pour charger les catégories Materiel from REST API
+  loadCategorieMateriel() {
+      this.categorieMaterielService.getAllCategorieMateriel().subscribe(
+  (bodyResponse) => { this.getCategorieMaterielData(bodyResponse); });
+  }
 
-getMessage(bodyResponse) {
-this.toasterService.pop('success', 'Supprimé avec succès!', bodyResponse.message);
-}
+  // get catégorie data from bodyresponse
+  getCategorieMaterielData(bodyResponse) {
+      this.categoriesMaterielsData = bodyResponse.data;
+      this.dtTrigger.next();
+  }
+  // delete catégorie
+  deleteCategorieMateriel (id) {
+    this.dataTableService.confirmDeleteMessage().then((result) => {
+      // delete the category from rest api
+      this.categorieMaterielService.deleteCategorieMateriel(id).subscribe(
+        (bodyResponse) => { this.getMessage(bodyResponse); });
+    }).catch(() => {});
+  }
+
+  getMessage(bodyResponse) {
+      this.toasterService.pop('success', 'Supprimé avec succès!', bodyResponse.message);
+      this.rerender();
+  }
+
+
+  ngOnDestroy(): void {
+    this.dtTrigger.unsubscribe();
+  }
+
+  rerender(): void {
+    this.dtElement.dtInstance.then((dtInstance: DataTables.Api) => {
+      // Destroy the table first
+      dtInstance.destroy();
+      // Call the rest api again
+      this.loadCategorieMateriel();
+    });
+  }
 }
